+++ packages/ui/src/components/Menu/index.ts
@@ -1,2 +1,4 @@
 export * from "./Menu";
-export * from "./Menu.types";
+export * from "./MenuItem";
+export * from "./MenuTrigger";
+export { useMenu } from "./useMenu";